Create Catalasegraph.py
This code is for the catalase graph of my lab for AP Bio, in this lab we investigated how substrate concentration, temperature, and surface area changed the rate of reaction for the decomposition of H2O2. In this graph, I include the class means for the rate of reaction using a piece of filter paper soaked in potato pulp, timing how long it took to rise to the top of the hydrogen peroxide solution. Standard error bars are included.